Moved Home.jsx to pages folder inside the home folder
